pyproject.toml
[project]
name = "npf-retail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "databricks-connect==16.1.0",
    "databricks-sdk>=0.65.0",
    "databricks-sql-connector>=4.1.2",
    "databricks-vectorsearch>=0.57",
    "ipykernel>=6.30.1",
    "pandas>=2.0.0",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
